Add badge alt text (#679)
- Add badge alternative text to Challenges, Triggers, Events, and Programs

+++ src/GRA.Domain.Service/BadgeService.cs
@@ -130,6 +130,7 @@ namespace GRA.Domain.Service
             badge.SiteId = GetCurrentSiteId();
             var result = await _badgeRepository.AddSaveAsync(GetClaimId(ClaimType.UserId), badge);
             result.Filename = await WriteBadgeFileAsync(result, imageFile, imageType: null);
+            result.AltText = badge.AltText?.Trim();
             return await _badgeRepository.UpdateSaveAsync(GetClaimId(ClaimType.UserId), result);
         }
 
@@ -139,29 +140,32 @@ namespace GRA.Domain.Service
         {
             var existingBadge = await _badgeRepository.GetByIdAsync(badge.Id);
 
-            if (File.Exists(existingBadge.Filename))
+            if (imageFile != null)
             {
-                File.Delete(existingBadge.Filename);
-            }
-
-            var imageType = ImageType.Png;
-
-            if (!string.IsNullOrEmpty(uploadFilename))
-            {
-                var uploadExtension = Path
-                    .GetExtension(uploadFilename)
-                    .ToLowerInvariant();
-
-                if (uploadExtension.EndsWith("jpg", StringComparison.OrdinalIgnoreCase)
-                    || uploadExtension.EndsWith("jpeg", StringComparison.OrdinalIgnoreCase))
+                if (File.Exists(existingBadge.Filename))
                 {
-                    imageType = ImageType.Jpg;
+                    File.Delete(existingBadge.Filename);
                 }
-            }
 
-            badge.Filename = await WriteBadgeFileAsync(existingBadge, imageFile, imageType);
-            return await _badgeRepository.UpdateSaveAsync(GetClaimId(ClaimType.UserId),
-                badge);
+                var imageType = ImageType.Png;
+
+                if (!string.IsNullOrEmpty(uploadFilename))
+                {
+                    var uploadExtension = Path
+                        .GetExtension(uploadFilename)
+                        .ToLowerInvariant();
+
+                    if (uploadExtension.EndsWith("jpg", StringComparison.OrdinalIgnoreCase)
+                        || uploadExtension.EndsWith("jpeg", StringComparison.OrdinalIgnoreCase))
+                    {
+                        imageType = ImageType.Jpg;
+                    }
+                }
+
+                badge.Filename = await WriteBadgeFileAsync(existingBadge, imageFile, imageType);
+            }
+            badge.AltText = badge.AltText?.Trim();
+            return await _badgeRepository.UpdateSaveAsync(GetClaimId(ClaimType.UserId), badge);
         }
 
         public async Task<Badge> GetByIdAsync(int badgeId)
